perf: reduce number of \WP_Query executions by combining tax queries
Previously, we looped through each category and tag term and executed a \WP_Query with tax_query for that single term. Now, we run a single query for all category terms and another for all tag terms. This should limit the number of tax queries to three at most. We could potentially query categories and tags together, but this seems to be the easiest way to retain knowledge of which posts in the results apply to which type of taxonomy.