Base plugin functionality (#5)
* feat: add plugin sidebar for handling sponsor meta

Adds a PluginSidebar to the Sponsor CPT only which allows editors to input and update sponsor metadata.

* feat: implement theme helpers to get sponsors for a post

Implements a collection of theme helper functions that allow theme developers to easily get an array of all sponsors associated with a post by post ID.

* fix: avoid PHP warnings when categories or tags are empty

Checks that the category and tag terms returned by WP are arrays before checking them for sponsorships.

* chore: add readme

* fix: missing translation for string

* perf: reduce number of \WP_Query executions by combining tax queries

Previously, we looped through each category and tag term and executed a \WP_Query with tax_query for that single term. Now, we run a single query for all category terms and another for all tag terms. This should limit the number of tax queries to three at most. We could potentially query categories and tags together, but this seems to be the easiest way to retain knowledge of which posts in the results apply to which type of taxonomy.

* perf: avoid tax query if there are no terms to query for

This improvement checks whether there are any terms to query for before executing the WP_Query.

* chore: update some old function documentation

* perf: use $post arg already supplied to wp_insert_post callback

* feat: new helper function to get sponsors for archive term

Creates a new helper function that can be called from an archive page to get sponsors for the archive's term.

* feat: add meta option to hide sponsor on posts if not direct

This new toggle option lets editors specify that particular sponsors should only be shown on posts if directly assigned. This lets sponsors appear on archive pages but not the posts that own that term.

* fix: add untracked editor SCSS file

* feat: move sponsor settings into Document settings sidebar

Moves the sponsor settings meta fields from a separate plugin sidebar to the existing Document settings sidebar. Also changes the Featured Image labels to read "Sponsor logo" instead.

* fix: do not remove theme editor styles

Remove a hook callback carried over from another plugin that removes theme editor styles.

* fix: re-add hook to remove Campaign settings from editor

* refactor: disallow specific editor assets

* feat: filter taxonomy sidebar panels to add message

Adds a message specific to Newspack Sponsors to the category and tag panels explaining what they do,

* chore: update function documentation